Add Saved Properties components & refactor
